+++ pkg/controller/controller_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/giantswarm/micrologger/microloggertest"
 	"github.com/prometheus/client_golang/prometheus"
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
 
 	"github.com/giantswarm/operatorkit/v4/pkg/resource"
@@ -162,6 +163,7 @@ func mustNewTestController(n string) *Controller {
 			Resources: []resource.Interface{
 				&testResource{},
 			},
+			Selector: labels.Everything(),
 
 			Name: n,
 		}